Return standard POS heading for Chinese Wiktionary template "=n="
Due to the strange syntax in the "NoEdit" template which is used in
"=n=", it can't be expaned correctly by wikitextprocessor. Return a
simplified heading so the pages use this template can be pasred.

The "=n=" template originally expands to `[[Category:{{{1}}}语名词
|{{{2}}}]]\n<div class=\"noedit\">\n===名词===\n</div>\n\n\n` but
wiktextract can't parse it as POS heading because the `<div>` tag.