Attempt to fix broken en.wiktionary.dump (20230120)
We had a typo in our Lua scribunto code, "untripNoWiki" instead
of "unstripNoWiki".

Mediawiki uses meta-characters called 'strip markers' that are
apparently similar to how we use magic characters, for marking
out NoWiki stuff etc.

Usually, these characters should not be expressed, but sometimes
they slip through and apparently the dump that was broken might
have been like that.

After fixing the typo (and removing the print statements so that
we don't get a log-file full of 140 000 000 prints of the same
two lines) things seemed to run well with tests.

Let's hope it works tomorrow!