Fixed broken parsing of HTML-tags
Fixes #207

For some unfathomable reason, when I introduced code in an earlier
commit that fixed and issue with parsing single quotes within
HTML-tags, I also broke parsing for HTML-tags.

Issue: a lone < in text was parsed as starting an HTML tag if
somewhere after it there was a >. In retrospect, quite bad.

Fix: create a more complete regex using ALLOWED_HTML_TAGS.

The rest was just debugging why re.split and the above regex would
strip obvious "</div>" tags into "div". It was completely baffling.

Turned out that in the new regex, it was capturing *all* the
capture groups, including the new one for
(different|html|tag|names), so the "stripped </div>" tag was
nothing of the sort, but a ghost of group(2) that was introduced
into the iteration stream.

This was corrected by making the group non-capturing with (?:...).