Ignore capitalization in HTML tags
A small addition to the previous commit which broke a test due to
it having capitalized HTML-tags.

The test was not caught before the weekend because I had done some
cleaning on my git branches and removed the `main` branch and left
`master` (which is from where?)... wiktextract has a `master`
branch instead of `main`, btw.