Changed regex to allow HTML-tags inside LINKs
Previously, parsing errors occurred when a [[link|xyz]] contained an
HTML tag: `[[link|<b>bar</b>]]`, because the regex used to detect
these for magical encoding (compacting into a magic character that
would be ignored in the meanwhile and expanded back again later when
appropriate) forbid lt and gt inside the link, which is actually not
forbidden in Wikitext.

This would cause garbage data and sometimes mess up the parsing of
outer templates (like #if-parser-functions) by breaking up the link
into separate pieces that were handled as two separate arguments.

Tatu was not sure why this was like this, but it is possible that it
was a fix to some other issue that might now pop up again; fingers
crossed.