Fixed issue with textual "=" in template arguments
Problem: if a template argument had a textual "=" in it, like in
quadratic/English's translation template, then that equal sign could
be passed forward inside the template argument and break other
templates and invokes downstream.

To fix this, I've put an extremely ugly kludge in a portion of the
code in core.py that handles expanding template arguments: if
there's a = in the argument and if that = is outside of any {{}}
template braces of <> html brackets, just... turn it into it's HTML
entity, which will eventually be transformed back at some point
(don't know when, but thank goodness I don't have to do it).