Adding parameters to ctx.process() for pre-expansion
Committing to get this out of the way for a while.

Added parameters to Wtp.process() that should in the future
contain sets of template names that should and should not be
expanded in the pre-expansion step when parsing the Wikitext
for the first time.

XXX Remember to remove the lang_code="zh" specific stuff to wiktextract,
where it logically belongs. Decouple wiktionaries-specific code
from wikitextprocessor and leave it agnostic.

XXX In wiktextract do all the stuff with the az-suffix-form template,
put the name in the do-not-expand set so that it can be extracted
later by using that name.