Add more decompositions of magical characters in string.gsub
Previously, only a few `%[` magical characters were decomposed so that
they work with gsub's replacement strings in our version of Lua.

This fix decomposes all magical characters I could google.

There are two replacement operations because Lua's pattern language does
not have an OR, so "start of string" and "not preceded by `%%`" could not
be in the same pattern.

I've left the stuff with "%-" alone, because it seems extra specific.