Two iterations of the escape decomposing needed
Fix to the previous commit; the fix was correct, and LUA errors fell
from 2.5M to around 90K. The final stragglers were instances of
overlapping patterns: say you have "%)%.", ")." in the original
unescaped text, and your pattern relies in looking at the character
before your match (in this case the percent+dot match): because the
percent+bracket has already "consumed" the pattern, it starts the next
search at the percent in percent+dot, which will not match.

This is simply fixed with a brute-force second round where we catch
any stragglers. I *think* this should take care of everything, because
the even in situations where you have three or more overlapping
patterns, you would do the odd-numbered matches in the first round and
the even-numbered matches in the second round.