Fixes more Lua replacement string errors
After committing the last patch, I realised there's a better
way to handle all of this. The previous code had grown a bit too much
organically, time to prune.

Currently, just replace `%%` (or `%%%%` when escaped...) with the
magic string `__PERC__` just for the duration of this operation.
With this, we don't have to worry about overlapping `%%[` or `%%%[`,
which also means we don't have to worry about matching at the
end of the string separate anymore.