Reverted unnecessary manipulation of Wtp.need_pre_expand
There was a much simpler path (in wiktextract) to have a blacklist
of templates not to expand, because it turned out that the pre-expand
step didn't actually affect the actual text cache at all, only
the Wtp.need_pre_expand set in the pickle. So you can just do
it on the fly as needed.