Quick and dirty fix to weird ! token behavior
{{tt}} templates broke our parser completely when there was a "!"
character at the start of an argument, like `|alt=!text`.

The ! was tokenized here, and thus table_hdr_cell_fn was called to
handle it when parsing it. However, the {{tt}} template somehow
managed to squeeze in a LINK node so that the ! token was inside it,
and NodeKind.LINK was not handled at all, which caused the parser
stack to pop a parse node at the end of the loop in
table_hdr_cell_fn. This caused basically everything to break in a
cascade of "not properly close" elements.

Quick and dirty fix: just added NodeKind.LINK to the list of node
kinds that should output the ! as text.