Insert `local arg = {...}` for Lua vararg
Lua's old vararg syntax up to 5.1 was that if there was a vararg parameter
`...` in a function signature (function name (parameters ...)), then those
extra parameters could be accessed through the hidden variable `arg`.
However, later this behavior was changed, and `arg` was replaced with
`...` everywhere, not just the signature; but `...` also behaves different
from `arg`, so you can't just substitute one for the other.

Happily, there seems to be (fingers crossed) a method to convert this
without massive code-changes.

What we do here: when a function signature has vararg `...`, detect it
with a regex (Lua function signature syntax is pretty simple), then insert
a line immediately after that that basically unpacks (and repacks) the
parameters in `...` into a local variable named `arg`. This way, our 5.4
Lua interpreted can handle `arg` identically (hopefully) to the old `arg`
behavior in 5.1, and nothing else needs to change because varargs are
local in scope to their function.