Workaround for a specific escape sequence mismatch
Because of the differences in escape sequences between Lua
5.1 and 5.4, something that will pass in 5.1 will not in
5.4, like "\/". "\/" is not an actual escape sequence in
either version, but 5.1 will fail it by just collapsing it
to "/" while 5.4 will crash.

Found in fr.wiktionary Module:locution, but added as an
escape sequence transformation to get it out of the way
before we move to a new Lupa version.