Branch: switch to luajit20 init
Create branch `lupa` for experimenting and testing changing
over to luajit20.

Lupa was cloned from github into a new python venv environment, with
--recurse-submodules, and these pip packages were installed:

- the cloned lupa repo's master
- nose2
- nose2[coverage_plugin]
- dateparser
- lru-dict

nose2 had to be installed into the venv to prevent the system's nose2
from just ignoring the packages in the venv, even if it was being
called from venv.

Might be some packages are still needed.

The tests were actually *hanging*, but it turned out to be a "sensible"
hang: in tests/test_wikiprocess > test_lua_lax_time1() we test to see
if the timeout feature of scribunto times out under ten seconds...
So disabling this test at least let the testing procedure end, with
many lua-related fails.

This initial commit is just to get the testing to actually run and
to initialize a new branch.