Some Lua 5.3 -> 5.1 changes
table.unpack()(5.3) -> unpack (5.1)

The weird "local global" _ENV variable does not exist in 5.1,
and should not be tested for.

Previously I was under the impression that the shift from
using `...` in function signatures and `arg` elsewhere to then
using `...` for 'both' happened between 5.1 -> 5.2, but it's
actually a 5.0 thing that was changed in 5.1, so we *shouldn't*
need to worry about it... But probably will.

In any case, disable syntax substitutions.