Kludge to handle italics inside table attributes
In translations you sometimes have translation tables with
italics or bold in their titles, which broke the table attribute
parsing (which was kind of simplistic, and still is, just kludged).

The italics and bold nodes are already part of the parse tree when
the attributes of a table are being checked, so the fix is just
to check whether the candidate "attribute line" is just made out
of strings and bold and italic nodes (recursively), and use
the flattened string of this delving if nothing else is detected.