Use regex to detect table attribute strings
Previously, table attribute strings were being detected
very loosely based on whether there was one singular contiguous
string in a node. This would break down due to weirdly expanded
and parsed nodes inside that same space created by things like
trans-top and multi-trans interleaving; so instead of

    `attr=1 attr2=2 attr3="&apos&apos&aposITALICS&apos&apos&apos"`

you'd get

    `attr=1 attr2=2 attr3="`, <ITALICS node>, `"\n`

which would break the parser and cause garbage output for the
"sense" from the table header.

This was previously fixed in the previous commit by flattening
text and italics/bold nodes, but this didn't actually fix
anything but those cases, and it turned out that this was
a fault somewhere deeper in the parser.

After being unable to figure out where the heck this thing is
breaking down and over-parsing things, I've just changed it so
that the attribute node candidate children nodes are all
turned back into wikitext, escaped and then checked with a
copy of the regex used later on to extract the attribute
assignment data. This seems to work.