Fixed issues with invalid escape characters
This took ages to figure out (because I got stuck on thinking that
LuaJIT had some kind of Unicode problem, but it was just my code
that was incomplete), although most of it was just the last mile.

LuaJIT does not behave identically to Lua 5.1 in regards to escaped
characters. The escape character tokens (\n, \t, etc.) are the same,
it's just that if you escape something that an invalid escape
character in Lua 5.1, it will silently ignore the backslash and
remove it, while LuaJIT will throw and error.

Even though all of the other substitutions in loader_replace_patterns
could be remove, this operation took its place to remove backslashes
that would have been ignored by Lua 5.1.

The final hurdle was realizing that backslashed digits are also
valid escaped characters. This lead to problems with string
manipulation, which I thought was an issue with Unicode, but
turned out to just be the unescaping of patterns like `[\128-\255]`
used by mw.ustring to search for Unicode characters in lua's
"regex-lite" patterns. The code for removing backslashes broke these,
causing mw.ustring to believe it could do certain things it couldn't,
borking some very specific gsub operations (rarely).

After fixing the above issue, it also magically fixed every other
error I had put on the backburner.