Explicitly disable test_lua_max_time1()
This test has been disabled as I waited to figure out other bugs first; it
would actually hang make test because it uses a simple infinite loop to test
the timeout-feature from Scribunto.

However, it turns out that the hangs were not caused by anything we did in
our code, but is instead because of the switch to using LuaJIT.

The timeout feature of Scribunto (or our code, it may actually be different
from Scribunto now that I think about it...) works by using debug hooks,
created with debug.sethooks that calls a debug function every now and then
(in our case, after every 100 000th instruction). The timeout hook just
checks to see if too much time has passed for that particular invokation of
Lua stuff.

The problem is that in LuaJIT, these debug hooks don't exist inside JITted
code.

So in our case, because our test loop was extremely simple (just i = i+1
forever), the JIT compiler did whatever a JIT compiler does and JIT compiled
the heck out of it, excluding the vital debug hook that was used by the
timeout feature.

This can be "fixed" by making the loop break out of the JIT (using a debug
print, for example, which seems to work fine), but it doesn't really change
the situation: in rare cases, timeout might just not work because it has been
disabled inside JIT code.

This commit is just to make it clear that this test is now disabled
"officially" and not just because I want to take a look at it later.