Temporary disable retaining Module:glossary in memory
Dump 20230401 from en.wiktionary broke Kaikki.org due to a bug in how
wikitextprocessor handles global names for retained (kept in memory) modules,
specifically Module:glossary.

This has been fixed by disabling retention for Module:glossary temporarily,
until a better fix can be made.

The issue:

Module:glossary had a small addition in the newest dump, in the form of a
function for formatting strings within that module.

That function was being called inside other functions that were being exported
out of the module. All well and good.

However, on our side this was broken, causing a bunch of Lua errors and
failure of the Kaikki.org regen process.

This was because the formatting function `format_anchor` was created as a
*global* name. If `format_anchor` is changed into a local function, it works.
Why? Because Lua's function encapsulation behaves differently under the hood
for local and global variables. Local variables are properly enclosed in
the function, they are passed into the function and kept there, while
global names are not; they have to be accessed through `_G`, the Lua
global namespace entity table.

`format_anchor` was present only once when calling Template:glossary on a page,
and every subsequent call of the Template:glossary (which calls Module:glossary)
failed because `format_anchor` was nil. Somewhere between the first successful
template call and the others, `format_anchor` was being cleared and never
re-created.

Module:glossary is a very high-frequency module that is called constantly.
To speed things up, there is a list of these kinds of modules that are
*retained*: they are not rerun or cleared from memory constantly, but
kept there. If we disable this retention, things start working again.

So, obvious in hindsight. The issue here is that we clear the global
namespace (and repopulate it with stuff again), including the globals
created by retained modules. So when the function tries to later
reach the global namespace for something that was created in a previous
call, it can't find it.