Enable the overriding of certain templates
See commit master 2a8fe38 in wiktextract for more about
the problem of templates that cause unparseable output
that need to be intercepted.

We introduce a new dict containing template-name and templatefn
pairs. When expand_recurse encounters one of these templates,
instead of handling the template through its own template code,
we use the template function instead.

The arguments of the template are expanded first, and then
handled to the template function, enabling recursive parsing.
For example, if you have egy-glyph-img inside an egy-glyph
template's arguments, that egy-glyph-img template is expanded
(and intercepted!) first.

Thus, instead of returning just an image (so no text data),
where egy-glyph-img was is now the hieroglyphic number code
used by egyptologists. This argument data is used in turn
by egy-glyph to construct a string that is pretty close
to being well-formatted <hiero></hiero> data (but not quite,
because we can't intercept things outside the template...)
which is inserted into the text without breaking everything
with tables and textless images.