Implement Lua 5.1 __tostring behavior
Apparently, in Lua 5.1 when the __tostring metamethod was called with
tostring(), the return value could be a table.

Some weird modules use this as a short-cut (VL-verb) to pass data with it.
However, in later versions of Lua this has, understandably, been
disallowed. __tostring should only return strings.

This has been implemented in our intercepting tostring() compatibility
function by checking if there is a __tostring field in the metatable of the
value received, and using that function instead of _orig_tostring() if so.