Fixed bug with Lua tostring changes
Forgot to check if metatable exists at all, and then
I didn't `make test`.

Well, turns out there's the exact right code for this
situation just sitting a few lines above!

Reminded about how and and or work in Lua (and some
other languages), they don't return boolean values
by default because nil is false and non-nil (except
for false) is true. So `and` returns the value of the
second part if the first is true and `or` returns
the value of the second part if the first one is
false. I think.