Add "\n" before every {{multiline}} template
When multiline is on the same line as trans-top, the parser's
partial parsing of lists breaks down because nothing is
allowed before the * in a list or list item.

To fix this, add a harmless (surely??) newline in front of
every multiline template.

XXX: do this with a template_fn from parse_translation_item instead!