Remove unefficient template name set oprations
`Wtp.expand()` only checks if a template should be expanded, so there
is not need to keep a huge template set in memory and only have to
check if a template exists or if it's in a smaller
`not_expand_templates` set. And creating all template names set plus
querying from database is slow as hell.