Treate URL in template argument as plain text
Fix `archivedate` argument not included error in
Module:quote(https://en.wiktionary.org/wiki/Module:quote). This error
is caused by the parser treats the URL argument as external link and
later converts the URL to wikitext link(`node_expand.to_html()`) in
the `[url]` form with breaks the Lua code.

Module:quote checks the URL argument, if the URL starts with
`https://web.archive.org/web/` then the `archivedate` argument is not
required. But since the arguemnt now starts with `[https://`, it
throws an error.