Name previously unnamed lambda functions for debugging
There's a bunch of errors having to do with the
lambda functions in call_set_functions() in luaexec.py,
but it would be nice to see which one is being a problem
(or all of them?), so here's an attempt to give them
names.

I suspect the bug is related to the xnt-decl bug, where a
module was (still is, the module was just fixed at source)
to polute the Lua engine's global namespace so that things
stopped working. Mainly this is because the bug is no
reproducable with only one page being parsed, you need
a proper batch, like the Kaikki regen.