Horrible kludge for package.loaders in Lua
`Module:ja-translit` is the only module in the Wiktionary
lua codebase that actually uses `package` for anything,
in this case a test using package.loaders (and the data
itself is loaded with a different function later? idgi).

`package.loaders` is a bit reduced in Scribunto and the other
entriesbesides package.loaders[2] (I assume) should not
be used.

With this special case in mind, we just add a new
`package` table in the Lua environment that has a
`loaders` TABLE (not a function) which should contain
different functions that would attempt to load stuff when
previous ones fail, but in our case we just populate it with
{nil, modloader}, where modloader returns the result of
new_loader.

It seems to work, and I am worried!