Explicitly name fields to SELECT, order by title
When getting pages, instead of *, SELECT by explicit
field names so that if we do changes in the database schema
nothing is messed up when using numbered indexes to
access those fields.

Pages queried for are now ordered by title so that we
have a more predictable ordering for debug purposes.