_finalize_expand: Empty `<nowiki>` elements -> `<nowiki/>`
Previously, `<nowiki>` elements would cause error messages
when they came from _finalize_expand, because they
"shouldn't" be there; but removing them is not 100%
the correct change, because an empty `<nowiki></nowiki>`
element is still meaningful and can break parsing of
wikitext entitites.

Fix: just output a `<nowiki/>` if the element is empty,
but remove the tags otherwise (the text has been escaped
already so it should be fine).