Handle MAGIC_NUMBER value generation with iterator
Instead of doing stuff like `MAGIC_NUMBER = MAGIC_BASE_NUMBER +10`
manually, it's a bit less cumbersome and errorprone to handle
things with a generator.

We use the private use area of Unicode for all sorts of magic
characters that we sprinkle in the text to replace text
and entities we don't want to handle at one time or another.

The start of the block has a few special characters with
specific meanings, which can be expanded by stealing from
the MAGIC_FIRST to MAGIC_LAST block at the end; those
characters are used for more dynamic entity-hiding during
parsing in the expansion phase, when we don't want to have
them in the way and creating misparsing issues.