close_begline_lists strengthened and used more often
If a LIST_ITEM contains a `<div>` that remains unclosed,
that div-tag will basically consume the rest of the document
so that nothing else is parseable after that.

This is because we do not check whether a new list item
should be part of a previous list (ie. whether the
parser stack should be popped until we reach the
correct level of depth). The div contains the rest of the
list, and the rest of the document.

THIS DOES NOT YET FIX THIS ISSUE. We do not check for
list depth within a list yet, BUT close_begline_lists()
does something similar as a function in other parsing
functions for tokens, and we can at least limit the damage
of an all-consuming unclosed entity by basically closing
lists (and the div that it contains) before the rest of
the document is consumed. Only the list is kind of wonky,
and needs to be fixed later. XXX

close_begline_lists also failed to correctly close an
open list when there were nodes in-between that weren't
NodeKind.LIST or NodeKind.LIST_ITEM (like HTML).

We now force it to check the parser stack and only
stop popping it when there are no more NodeKind.LIST
layers left.