Add type-checking dependency `type-lxml` and create typestubs dir
It would be nice, if possible, to have stubs for Lupa,
so I'm pre-emptively adding a directory to work in.

However, stubgen generates an empty __init__.pyi file
with only comments with un-generated entities. Lupa is
written in Cython *and* it involves passing Lua and
Python objects back and forth, so it's a bit messy.