Fix assert causing exceptions, fixed num_threads bug
Turns out assert will fail on empty strings, too,
not just None. Oops. Replaced with an if block.
Core.py needs a refactor, it's getting tight in
the margins.

I previously changed the init default value of
Wtp.num_threads = 1, but this was a mistake;
None is a perfectly valid value here, because
multiprocessing.Pool() interpretes it as "all
possible processes".

However, this didn't cause any issues when running
wiktwords, because it gets explicitly set to None when checking
command line arguments in the section with argparser.