Disable checking of ctx.beginning_of_line in magic_fn
magic_fn() handles previously preparsed and then encoded
entities like templates, triple-curlies, urls and links,
which have arguments that use `|` vbars as separators.

When looping over these arguments, the vbars are inserted
in-between them using their parsing function vbar_fn, which
caused misparsing because then those bars would be parsed
wrongly if they were found a the beginning of a line.

To stop this, ctx.beginning_of_line was reset to False when
looping through the arguments in templates.

I found later issues with this same thing with the article
`0800 number`, where you had a template like
```
{{foo
|bar=
[baz]
}}
```

and then failed to fix this with a simple solution.

This patch creates a new 'static' context-manager-like object
called ctx.begline_disabled, initialized in Wtp.__init__, which
does two things: on entering a with context, it increases a counter
and flips ctx.begline_enabled to False, and on exciting it
decreases the counter and checks to see if it's <= 0, in which
case it enables ctx.begline_enabled again.

Anywhere ctx.beginning_of_line is being checked there is also
an added `and ctx.begline_enabled` that will falsify the condition
when set.

ctx.begline_disabled can be used with `with`:

```python
with ctx.begline_disabled
    ...
```