Edited and commented regexes.
`{` and `{` need to be escaped in regexes only
rarely, when otherwise they would be parsed as
`{1}, {1,2} or \N{UNICODE NAMED} (Python 3.8).

Otherwise, I mostly just added newlines and
whitespace for legibility, plus comments.

One change: in one instance of finding `{|..|}`
the inner contents was being consumed greedily,
so changed `*` into `*?`.