Changes to Lupa.lua51 stub file, replaced __init__.pyi
This stub file now contains a pretty comprehensive type stub
for lupa.lua51.

The original was created by using stubgen on *something*, a
C-source file somewhere (in the pip installed library files),
not the Cython source which can't be parsed by stubgen... But
I can't recreate it! So the original data has been left at the very
end of the stubfile, gathered from an __init__ file.

I then expanded on what I could see in the initial stubgen-file,
trying to recreate the signatures based on the Cython code
in the Lupa source.