Set `need_pre_expand` if a template is redirected from a pre expand template
For example: Template:a redirects to Template:b,
`Wtp.analyze_templates()` only set page "a" `need_pre_expand` to
`True` if page "b" needs pre expand.

This commit sets `need_pre_expand` to `True` for page "b" if page "a"
needs pre expand.

Also set `need_pre_expand` to `True` if a Chinese Wiktionary templates
uses the `langhd` template becasue `langhd` is used to create language
section subtitles.

And make the Chinese Wiktionary language and POS subtitle
template check more strick, now the code only set need_pre_expand to
true if a template name starts with `-` or `=` but also ends with `-`
or `=`, because Template:- is not a subtitle template.