More type-checking stuff, luaexec.py
I am doing luaexec out of the way because it's the most
annoying part to type-check, because there is no way
to integrate the type-checking with the Lua stuff other
than by brute force.

There's a couple of
```
ctx.NAMESPACE_DATA
    .get("Template", {"id": None})
    .get("id"))
```
style monstrosities I modified out of the original,
because the typechecking with TypedDictionaries
are so finnicky; I *know* the {"id": None} could just
be {}, but we're initializing a dictionary anyhow,
and putting the "id" in there makes mypy shut up.

Technically it could done with a final `["id"]`,
because we can pretty much trust NAMESPACE_DATA,
but just in case we get something weird from
NAMESPACE_DATA a `.get("id")` is a final sentinel.