Trying out regex re.X flag: ignore whitespace and comments
The re.X flag (`(?x)` to its friends) can be added to
re.compile() and other re-module functions that take
regex patterns as a bitwise-or'ed flag in `flags=`.

Its function is to make the regex parser ignore whitespace
in the regex pattern and allow python-style #comments at
the end of a physical line.

It's purpose is to allow regex-writers to write more easily
readable patterns, without mixing different strings and
python syntax into it.

This is just a test to see how things work out.