_lua_invoke / ctx.lua_invoke should return string
This is rare that it happens in practice, but #invoke calls
to functions (which _lua_invoke/ctx.lua_invoke is for)
should return strings.

According to the documentation, return values will be stringified
and "concatenated", but testing it out on Wiktionary's sandbox shows
that tables and functions return the string "table" and "function"
respectivelly, `nil` values are returned as empty strings, and
the rest are stringified as normal.

The differences might be because the stringification happens in
PHP instead of inside Lua. Couldn't find any __toString methods
in the source code, so really, no idea what's happening there.