Bug fix, added `with` multiprocessing context manager, type annotations
Stupid bug that didn't break tests but broke wiktwords now fixed.
`if t is None` reverted, don't know what I was braining.

Added a with-block around the bits with multiprocessing.pool.Pool()
in core.py. The return object `pool` from multiprocessing.Pool()
is actually also a context-manager, so you can use it with a
`with` to create a block that always calls pool.terminate() at the end.

Nevermind that we weren't calling pool.terminate() before! Just
pool.close() and pool.join(), which we still have to do manually,
because those are not in __exit__(). But its better to have the
context manager still close the pool in case an error occurs, not
that it really matters in our case because the whole program
is FUBAR'd if anyhow, this is just good practice. I hope.