Do not append `...#:` 'continuation' lines to parent
https://github.com/tatuylonen/wikitextprocessor/issues/84

Currently we are following Wikitext behavior stricly, which means
that in lists, lines with `...#:` style signatures are actually
'continuations' of the parent node, if there are intervening
sublists.

This causes some problems with detecting whether certain items
are on their own line, because when appending the items that
information is lost.

Because we can reasonably easily detect whether a "...#:" should
be appended into the parent's contents after we have parsed
the page, leaving it to the user how to interpret this seems
reasonable. Currently, for Wiktionary purposes it is best
to just leave the List and List Item wikinodes, because that
data is used in certain places (like when determining what is
an example in a list of glosses).

No data should be "lost", unlike currently.