Fixed bug in parser.py/tags_fn where we didn't close lists
I am an idiot, and here's two lines of code that I didn't add the first
time I should have, and then remained completely blind to later on.

All the basic tests passed, and the code in the commit that this
one fixes did handle an issue we had, it just slipped through because
we hadn't come across examples where it broke.

Recently we had issues with 'frozen' (just exponential) processes
that stopped kaikki from updating. The issue was ultimately that
a template creating an HTML Div was put into the parse tree as part
of a LIST_ITEM, and because some of those template divs (synonym
table generating template) were BIG, when passed eventually into wiktextract
it would try to parse the whole div in places like decode_tags as
one big long set of tags.

The algorithm in decode_tags (and associated bits of code)
is mildly exponential, so when we come across a big, big set of tags
to decode, the time and memory requirements got out of hand and
kaikki would basically hang, with many 'hanging' processes.