Set `Wtp.get_page()` cache size to 1000
The English Wiktionary has 40327 template and 54524 module
pages(exclude redirects). I think the cache is mostly useful for
caching shared templat and module pages used in pages processes in a
worker process. So the cache size relates to page numbers, CPU core
numbers and shared templates. If 1000 is not enough, we can increase
it to 10000.