Use varargs in Lua wrapper function
This is such a gosh darned kludge, but... When
parsing Wikipedia wikitext, there's a module that
checks whether the getParent field of something is
actually a function (I don't know why).

The problem is that the getParent field in our very
specific case is actually a Python function passed
into Lua code through Lupa, which means its actual
type() return value is not 'function' but 'userdata'.

This can be worked around by creating a wrapper
function in Lua code that takes the arguments passed into
it and then calls the Python-function-inside-Lua.
The wrapper's type() is 'function' because it's a
native Lua function, and the Wikipedia module doesn't
complain.

However, we have a bunch of print statements inside
various Python-in-Lua functions that catch it
when the function is passed too many arguments, that
is, the bug where a few of these functions swap places
/ their references get scrambled in weird ways.
It would be nice to still have these debug prints
for getParent (which gets swapped every now and then),
so we need to pass through all the arguments form
Lua into debugGetParent using Lua's `...` varargs
syntax.