Do not attempt to parse templates with a first argument
Example: amante 20230829, Spanish has a template
```
{{|es|...}}
```
which caused some breakage in a new bit of code that
tried to get a slice out of the first argument, ''.

Wikimedia ignores templates like these and outputs
them as text, so the most forward way to handle this
is return the text in repl_templ where it would otherwise
cookie-ize the template into a special magic character
to be expanded later.

repl_templ is used by a re.sub() call as its "how to handle
this piece of substution" handler function, so this
should just leave the original piece of text un-touched.w