Test for closing HTML tags in list items.
Just realized I put a test for this in Wiktextract, when
it really should have gone here, and in a simpler form.

This is only a partial test for a partial feature, because
the parser doesn't fulfil the requirements completely:

When inside a list item, an element should be able to contain
newlines, as much as it pleases. Unfortunately, the parser
we have isn't capable of that (yet? XXX), but I was able
to kludge together a special case for when you have a
</end tag> that starts on a new line just after a list item:

```
* Item 1 <strong>foo
</strong>
* Item 2
```

This is all one list, and the whole `<strong>foo\n</strong>`
element is part of the first LIST_ITEM node.