Use `rowid` in addition to `title` in get_all_pages
This is just in case, but I kind of need a deterministic
ordering of get_all_pages output for testing, and
`ORDER BY title, rowid` should be 100% the same each time.
Just sorting by title could lead to (might not, I'm not sure)
different orders when there are several identical titles.