Wikitext outside URLs with [...] do not use vbar-arguments
Fixes https://github.com/tatuylonen/wiktextract/issues/321

Wikitext urls in the form [https://test.com] or [http://test.com TEST]
do not use |-vbars as separators for arguments. Documentation
indicates that you shouldn't use |-vbars that are unescaped
in urls, but it doesn't actually break the Wikimedia parser.

It did break ours, because there was an assumption that because
all other WikiNodes that were being parsed used |-separators,
URLs did too (or there was a one-line oversight). The fix here
is just to make an exception when parsing |-tokens for URLs.