Revert `?` added to the analyze template table regex pattern
This extra `?` causes catastrophic backtracking in page
https://en.wiktionary.org/wiki/Template:sw-conj/table

But the regex still has more backtrack bugs if it can't match a table,
we're just lucky the templates are not long enough to halt the
command. Many templates are marked need pre-expand is because the
table can't be matched.

We should use a simpler approach to count and compare the numbers of
`{|` and `|}`.