Add newlines before key chars when expanding
https://meta.wikimedia.org/wiki/Help:Newlines_and_spaces#Automatic_newline

Before certain key character, "*", ";", ":", "{|", "#",
wikitext expansion (afaict both templates and parserfns)
should have an inserted newline '\n' at the start of
the string.

This doesn't directly fix the original issue where this
was detected, but this small commit should at least shift
the code to be more correct.

There may be kludges still left in other parts of the code
that emulate this behavior for specific contexts, but
those probably rely on heuristically detecting when something
is wrong, so they shouldn't by mistriggered easily by this
change.

If some code is found that inserts newlines or splits lines
before the characters mentioned above, it might be ok to
remove that kludge.