Remove code added from commit 6d2d466b2a640961c68cb93db6330f40cdf8a588
I'm not sure what error that commit was trying to fix, the current
code has no problem process the translation section of the
"quadratic" page and the "trans-top" template's body is
`{{#invoke:translations|top}}`. Even if a template parameter is passed
to `#invoke` and it contains `=` the code is still correct and don't
need to replace `=`.

Remove this code fixes the Lua error in page "can" because `=` is
replaced to `&#61;` and causes a error in module "translations/multi".