Lowercase type name for Python 3.9+
"Dict[something, something]" -> "dict[something, something]"

In Python 3.9+ (probably earlier, but this is the relevant
version here) you don't need to import `Dict` and `List` and
other container type-hinting entities, but can use just the
same keyword as when creating or casting objects.

`Union` and `Optional` (unless in the form `Union[None, something]`
are still mandatory for 3.9, but will go away in the far future.